#523 refactor main bootstrap logic
